Add ActorSensorUtil.h
