+++ lib/al/Project/HitSensor/HitSensorDirector.cpp
@@ -2,7 +2,7 @@
 
 #include "Library/Execute/ExecuteDirector.h"
 #include "Library/Execute/ExecuteTableHolderUpdate.h"
-#include "Library/LiveActor/ActorSensorFunction.h"
+#include "Library/LiveActor/ActorSensorUtil.h"
 #include "Project/HitSensor/HitSensor.h"
 #include "Project/HitSensor/SensorHitGroup.h"
